refactored decorators to dedicated module
